Renamed and moved HTML-class to utilities
