+++ src/main/java/servlets/HelloServlet.java
@@ -1,6 +1,6 @@
 package servlets;
 
-import html.Html;
+import utilities.HTML;
 import java.io.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
@@ -13,7 +13,7 @@ public class HelloServlet extends HttpServlet {
 
         // Hello
         PrintWriter out = response.getWriter();
-        Html.Start(out, "Hello World! Servlet");
+        HTML.Start(out, "Hello World! Servlet");
         out.println("<h2>Search for a player's name in the database :-)</h2>");
         out.println("<form action='GetPlayerServlet' method='GET'>");
         out.println("  <label for='pname'>Nickname:</label>");
@@ -21,8 +21,8 @@ public class HelloServlet extends HttpServlet {
         out.println("  <input type='submit' />");
         out.println("</form>");
         out.println("<p>This query finds a player in your private Battlefield 2 statistics database and lists up some of the player's stats (I have not included every player column in the model.)</p>");
-        Html.MiniTutorial(out);
-        Html.EndBasic(out);
+        HTML.MiniTutorial(out);
+        HTML.EndBasic(out);
     }
 
     public void destroy() {
